Improved, cleaned and origanized upload center code
